+++ crate/operator/config.py
@@ -153,6 +153,9 @@ class Config:
     #: From which version onwards CrateDB supports JWT authentication
     CRATEDB_JWT_AUTH_VERSION: str = "5.7.2"
 
+    #: From which version onwards CrateDB supports global JWT config
+    CRATEDB_JWT_GLOBAL_CONFIG_VERSION: str = "5.10.1"
+
     #: Interval in seconds for which the operator will ping CrateDBs for their
     #: current health.
     CRATEDB_STATUS_CHECK_INTERVAL: Optional[int] = 60